minor fixes and refactor repo with interactors
