+++ app/src/main/java/ru/prometeydev/movie/App.kt
@@ -3,10 +3,7 @@ package ru.prometeydev.movie
 import android.app.Application
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
-import ru.prometeydev.movie.di.databaseModule
-import ru.prometeydev.movie.di.networkModule
-import ru.prometeydev.movie.di.repoModule
-import ru.prometeydev.movie.di.viewModelModule
+import ru.prometeydev.movie.di.*
 
 class App : Application() {
 
@@ -15,7 +12,7 @@ class App : Application() {
 
         startKoin {
             androidContext(this@App)
-            modules(listOf(networkModule, repoModule, viewModelModule, databaseModule))
+            modules(listOf(networkModule, repoModule, viewModelModule, databaseModule, notifyModule))
         }
     }
