+++ app/src/main/java/ru/prometeydev/movie/di/NotificationsModule.kt
@@ -1,0 +1,8 @@
+package ru.prometeydev.movie.di
+
+import org.koin.dsl.module
+import ru.prometeydev.movie.ui.MovieNotifications
+
+val notifyModule = module {
+    single { MovieNotifications(context = get()) }
+}
