+++ app/src/main/java/ru/prometeydev/movie/di/RepositoryModule.kt
@@ -1,10 +1,12 @@
 package ru.prometeydev.movie.di
 
 import org.koin.dsl.module
-import ru.prometeydev.movie.model.MoviesRepository
-import ru.prometeydev.movie.ui.MovieNotifications
+import ru.prometeydev.movie.model.Repo
+import ru.prometeydev.movie.model.usecases.MoviesInteractor
+import ru.prometeydev.movie.model.usecases.TopRatedMovieInteractor
 
 val repoModule = module {
-    single { MoviesRepository(api = get(), db = get(), notifications = get()) }
-    single { MovieNotifications(context = get()) }
+    single { Repo(api = get(), db = get()) }
+    single { MoviesInteractor() }
+    single { TopRatedMovieInteractor() }
 }
