+++ app/src/main/java/ru/prometeydev/movie/model/Repo.kt
@@ -1,0 +1,6 @@
+package ru.prometeydev.movie.model
+
+import ru.prometeydev.movie.model.database.MoviesDatabase
+import ru.prometeydev.movie.model.network.MoviesApi
+
+class Repo(val api: MoviesApi, val db: MoviesDatabase)
