+++ app/src/main/java/ru/prometeydev/movie/model/database/MoviesDatabase.kt
@@ -20,7 +20,7 @@ import ru.prometeydev.movie.model.database.entitiy.MoviesRemoteKeysEntity
         MovieEntity::class,
         MoviesRemoteKeysEntity::class
     ],
-    version = 9,
+    version = 10,
     exportSchema = false
 )
 @TypeConverters(DataConverter::class)
