+++ app/src/main/java/ru/prometeydev/movie/model/domain/Movie.kt
@@ -13,5 +13,5 @@ data class Movie(
         val runtime: Int? = null,
         val genres: List<Genre> = emptyList(),
         val actors: List<Actor> = emptyList(),
-        val video: String
+        val video: String? = null
 )
