+++ app/src/main/java/ru/prometeydev/movie/ui/MainActivity.kt
@@ -24,9 +24,9 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    override fun onNewIntent(intent: Intent?) {
+    override fun onNewIntent(intent: Intent) {
         super.onNewIntent(intent)
-        intent?.let(::handleIntent)
+        handleIntent(intent)
     }
 
     private fun handleIntent(intent: Intent) {
